[qatl] divided managers classes to own files
