+++ qatestlink/core/testlink_manager.py
@@ -2,15 +2,11 @@
 """Testlink Managers"""
 
 
+from qatestlink.core.utils.Utils import read_file
 from qatestlink.core.utils.logger_manager import LoggerManager
 from qatestlink.core.connections.connection_base import ConnectionBase
-from qatestlink.core.utils.Utils import read_file
-from qatestlink.core.xmls.route_type import RouteType
-from qatestlink.core.xmls.error_handler import ErrorHandler
-from qatestlink.core.xmls.request_handler import RequestHandler
-from qatestlink.core.xmls.response_handler import ResponseHandler
-from qatestlink.core.xmls.base_handler import BaseHandler
-from qatestlink.core.models.tl_models import TProject
+from qatestlink.core.xmls.xmlrpc_manager import XMLRPCManager
+
 
 
 PATH_CONFIG = 'qatestlink/configs/settings.json'
@@ -84,10 +80,10 @@ class TLManager(object):
         """Call to method named 'tl.getProjects'"""
         if dev_key is None:
             dev_key = self._settings.get('dev_key')
-        req_data = self._xml_manager.req_get_tprojects(dev_key)
+        req_data = self._xml_manager.req_tprojects(dev_key)
         res = self._conn.post(self._xml_manager.headers, req_data)
         self._xml_manager.parse_errors(res.text)
-        res_as_models = self._xml_manager.res_get_tprojects(
+        res_as_models = self._xml_manager.res_tprojects(
             res.status_code, res.text, as_models=True)
         # TODO: filter by name and/or value
         return res_as_models
@@ -96,150 +92,23 @@ class TLManager(object):
         """Call to method named 'tl.getTestProjectByName'"""
         if dev_key is None:
             dev_key = self._settings.get('dev_key')
-        req_data = self._xml_manager.req_get_tproject_by_name(
+        req_data = self._xml_manager.req_tproject_by_name(
             dev_key, tproject_name)
         res = self._conn.post(self._xml_manager.headers, req_data)
         self._xml_manager.parse_errors(res.text)
-        res_as_model = self._xml_manager.res_get_tproject_by_name(
+        res_as_model = self._xml_manager.res_tproject_by_name(
             res.status_code, res.text, as_model=True)
         return res_as_model
 
-
-class XMLRPCManager(object):
-    """
-    Manage all XMLRPCManager requests,
-     responses and handle errors. This class
-     store all official methods names used
-     on original XMLRPC php class
-    """
-    _request_handler = None
-    _response_handler = None
-    _error_handler = None
-
-    log = None
-    headers = None
-    handler = None
-
-    def __init__(self, log):
-        self.log = log
-        self._request_handler = RequestHandler(self.log)
-        self._response_handler = ResponseHandler(self.log)
-        self._error_handler = ErrorHandler(self.log)
-        self.headers = {'Content-Type': 'application/xml'}
-        self.handler = BaseHandler(self.log)
-
-    def parse_errors(self, xml_str):
-        """Raise an exception if response have error structure"""
-        #TODO: make enum and custom exception for each exception number
-        self._error_handler.get_response_error(xml_str)
-
-    def req_check_dev_key(self, dev_key):
-        """
-        :return:
-            string xml object ready to use on API call
-        """
-        req = self._request_handler.create(
-            RouteType.TLINK_CHECK_DEV_KEY)
-        return self._request_handler.create_param(
-            req, 'struct', 'devKey', dev_key)
-
-    def res_check_dev_key(self, status_code, res_str):
-        """
-        :return:
-            string xml object ready to
-             parse/write/find/add Elements on it
-        """
-        if status_code != 200:
-            raise Exception(
-                "status_code invalid: code={}".format(
-                    status_code))
-        return self._response_handler.create(
-            RouteType.TLINK_CHECK_DEV_KEY, res_str)
-
-    def req_get_tprojects(self, dev_key):
-        """
-        Obtains all test projects created on remote
-         testlink database, 
-         TODO: can filter with any property+value combination
-
-        :return:
-            List of TProject objects containing all database
-             data loaded
-        """
-        req = self._request_handler.create(
-            RouteType.TPROJECTS)
-        return self._request_handler.create_param(
-            req, 'struct', 'devKey', dev_key)
-
-    def res_get_tprojects(self, status_code, res_str, as_models=True):
-        """
-        Parse and validate response for method
-         named 'tl.getProjects', by default response list
-         of TProject objects, can response xml string too
-        :return:
-            if as_models is True
-                list of objects instanced with
-                 Model classes
-            if as_models is False
-                string xml object ready to
-                 parse/write/find/add Elements on it
-        """
-        if status_code != 200:
-            raise Exception(
-                "status_code invalid: code={}".format(
-                    status_code))
-        res = self._response_handler.create(
-            RouteType.TPROJECTS, res_str)
-        if not as_models:
-            return res
-        res_members_list = self._response_handler.get_response_members(
-            xml_str=res)
-        tprojects = list()
-        for res_members in res_members_list:
-            tproject = TProject(res_members)
-            tprojects.append(tproject)
-        return tprojects
-
-
-    def req_get_tproject_by_name(self, dev_key, tproject_name):
-        """
-        Obtains all test projects created on remote
-         testlink database, can filter by name
-
-        :return:
-            TProject object containing all database
-             data loaded
-        """
-        if tproject_name is None:
-            raise Exception("Can't call XMLRPC without param, tproject_name")
-        req = self._request_handler.create(
-            RouteType.TPROJECT_BY_NAME)
-        req = self._request_handler.create_param(
-            req, 'struct', 'devKey', dev_key)
-        req = self._request_handler.add_param(
-            req, 'testprojectname', tproject_name)
-        return req
-
-    def res_get_tproject_by_name(self, status_code, res_str, as_model=True):
-        """
-        Parse and validate response for method
-         named 'tl.getTestProjectByName', by default response
-         TProject object, can response xml string too
-        :return:
-            if as_models is True
-                object instanced with Model classes
-            if as_models is False
-                string xml object ready to
-                 parse/write/find/add Elements on it
-        """
-        if status_code != 200:
-            raise Exception(
-                "status_code invalid: code={}".format(
-                    status_code))
-        res = self._response_handler.create(
-            RouteType.TPROJECT_BY_NAME, res_str)
-        if not as_model:
-            return res
-        res_members_list = self._response_handler.get_response_struct_members(
-            xml_str=res)
-        return TProject(res_members_list)
+    def api_tproject_tplans(self, tproject_id, dev_key=None):
+        """Call to method named 'tl.getProjectTestPlans'"""
+        if dev_key is None:
+            dev_key = self._settings.get('dev_key')
+        req_data = self._xml_manager.req_tproject_tplans(
+            dev_key, tproject_id)
+        res = self._conn.post(self._xml_manager.headers, req_data)
+        self._xml_manager.parse_errors(res.text)
+        # TODO: make works
+        res_as_models = self._xml_manager.res_tproject_tplans(
+            res.status_code, res.text, as_models=True)
+        return res_as_models
